Move UpdateRunModel to its own file
