+++ src/ert/run_models/ensemble_smoother.py
@@ -11,6 +11,7 @@ from pydantic import PrivateAttr
 from ert.config.parsing.config_errors import ConfigValidationError
 from ert.enkf_main import sample_prior, save_design_matrix_to_ensemble
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.run_models.update_run_model import UpdateRunModel
 from ert.trace import tracer
 
 from ..analysis import smoother_update
@@ -18,7 +19,7 @@ from ..config import DesignMatrix, ParameterConfig, ResponseConfig
 from ..plugins import PostExperimentFixtures, PreExperimentFixtures
 from ..run_arg import create_run_arguments
 from ..storage import Ensemble
-from .base_run_model import ErtRunError, UpdateRunModel
+from .base_run_model import ErtRunError
 
 logger = logging.getLogger(__name__)
