+++ src/ert/run_models/multiple_data_assimilation.py
@@ -17,13 +17,14 @@ from ert.config import (
 )
 from ert.enkf_main import sample_prior, save_design_matrix_to_ensemble
 from ert.ensemble_evaluator import EvaluatorServerConfig
+from ert.run_models.update_run_model import UpdateRunModel
 from ert.storage import Ensemble
 from ert.trace import tracer
 
 from ..analysis import smoother_update
 from ..plugins import PostExperimentFixtures, PreExperimentFixtures
 from ..run_arg import create_run_arguments
-from .base_run_model import ErtRunError, UpdateRunModel
+from .base_run_model import ErtRunError
 
 logger = logging.getLogger(__name__)
